Create circuit compiler stub and reformat some code
