+++ src/core/compiler/dfs.rs
