+++ src/core/mod.rs
@@ -1,2 +1,3 @@
 pub mod canvas;
 pub mod simulation;
+pub mod compiler;
